beta image getter working on template 2
