+++ public/index.html
@@ -3,6 +3,7 @@
   <head>
     <title>Capstone</title>
     <script src="/bundle.js" defer></script>
+    <script src="https://upload-widget.cloudinary.com/global/all.js" type="text/javascript"></script>
     <link rel="stylesheet" href="/main.css">
     <meta name="viewport" content="width=device-width, initial-scale=1">
   </head>
