+++ src/components/student/_BookEditor.js
@@ -66,7 +66,7 @@ const BookEditor = () => {
                     </div>
 
                     <div className="page-editor">
-                        <templates.Template3 page={currentPage} />
+                        <templates.Template2 page={currentPage} />
                     </div>
 
                     <PublishDeleteButtons bookId={currentBook?.id} />
